+++ rtsim/src/gen/site.rs
@@ -32,6 +32,7 @@ impl Site {
             | SiteKind::ChapelSite(_)
             | SiteKind::Terracotta(_)
             | SiteKind::Gnarling(_)
+            | SiteKind::Cultist(_)
             | SiteKind::PirateHideout(_)
             | SiteKind::JungleRuin(_)
             | SiteKind::RockCircle(_)