+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added reworked Cultist Dungeon
 - Petting animals tamed by you or someone else!
 - Updated furniture sprites.
 - Added Abyssal rings